+++ src/com/finVault/model/Transaction.java
@@ -5,22 +5,23 @@ import com.finVault.enums.TransactionStatus;
 import com.finVault.enums.TransactionType;
 import com.finVault.utils.IDGenerator;
 
+import java.math.BigDecimal;
 import java.time.LocalDateTime;
 
 public class Transaction extends BaseEntity {
     private final String transactionId;
     private final TransactionType transactionType;
-    private final double amount;
+    private final BigDecimal amount;
     private final TransactionStatus transactionStatus;
     private final LocalDateTime transactionDateTime;
     private final String fromAccountNumber;
     private final String toAccountNumber;
-    private final double balanceAfterTransaction;
+    private final BigDecimal balanceAfterTransaction;
     private final String transactionDescription;
 
-    public Transaction(TransactionType transactionType,double amount,
+    public Transaction(TransactionType transactionType,BigDecimal amount,
                        String fromAccountNumber, String toAccountNumber, String createdBy,
-                       double balanceAfterTransaction, TransactionStatus transactionStatus,
+                       BigDecimal balanceAfterTransaction, TransactionStatus transactionStatus,
                        String transactionDescription) {
         LocalDateTime localDateTime=LocalDateTime.now();
         this.transactionId = IDGenerator.generateTransactionId();
@@ -37,12 +38,12 @@ public class Transaction extends BaseEntity {
     }
 
     public String getTransactionId() { return transactionId; }
-    public double getAmount() { return amount; }
+    public BigDecimal getAmount() { return amount; }
     public String getFromAccountNumber() { return fromAccountNumber; }
     public String getToAccountNumber() { return toAccountNumber; }
     public TransactionType getTransactionType() { return transactionType; }
     public LocalDateTime getTransactionDateTime() { return transactionDateTime; }
-    public double getBalanceAfterTransaction() { return balanceAfterTransaction; }
+    public BigDecimal getBalanceAfterTransaction() { return balanceAfterTransaction; }
     public String getTransactionDescription() { return transactionDescription; }
     public TransactionStatus getTransactionStatus() { return transactionStatus; }
 }
